requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anndata
array-api-compat
asciitree==0.3.3
asttokens==3.0.0
attrs==25.3.0
Bottleneck
Brotli
brotlicffi
cached-property
certifi
cffi
charset-normalizer
colorama
comm==0.2.3
contourpy
cycler
debugpy==1.8.17
decorator==5.2.1
Deprecated
einops==0.8.1
executing==2.2.1
fasteners
filelock
fonttools
frozenlist==1.7.0
fsspec
gmpy2
h5py
hf-xet==1.1.10
huggingface-hub==0.35.0
idna
igraph
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2
joblib
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver
legacy-api-wrap
leidenalg
lightning-utilities
llvmlite==0.44.0
MarkupSafe
matplotlib==3.10.6
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
msgpack
multidict==6.6.4
natsort
nest-asyncio==1.6.0
networkx
numba
numcodecs
numexpr
numpy
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging
pandas==2.0.3
parso==0.8.5
patsy
pexpect==4.9.0
pillow
pkgconfig
platformdirs==4.4.0
ply
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser
Pygments==2.19.2
pynndescent
pyparsing
PyQt5==5.15.11
PyQt5_sip==12.17.0
PyQt6_sip
PySocks
python-dateutil
pytorch-lightning
pytz
pyvips
PyYAML
pyzmq==27.1.0
requests
safetensors==0.6.2
scanpy
scikit-learn
scipy
scprep==1.2.3
seaborn
session-info2
setuptools==78.1.1
sip
six
stack-data==0.6.3
statannotations==0.7.2
statsmodels
sympy
texttable
threadpoolctl
tifffile
timm==1.0.19
toml
tomli
torch==2.4.0
torchmetrics
torchvision==0.19.0
tornado
tqdm
traitlets==5.14.3
triton==3.4.0
typing_extensions
tzdata
umap-learn
unicodedata2
urllib3
wcwidth==0.2.13
wheel==0.45.1
wrapt
yarl==1.20.1
zarr
